requirements.txt
streamlit>=1.28.0
groq>=0.4.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-groq>=0.1.0
pandas>=2.2.0
plotly>=5.17.0
requests>=2.31.0
pydantic>=2.0.0
openai>=1.0.0